Revert "WIP: Use actual modules for the ember extractors -- however, formatjs is"
This reverts commit 57d64739d7940d210938bcbd81ae2548d7a565cf.

Have CI tell me what command was failaing
