+++ package.json
@@ -43,9 +43,9 @@
     "@lerna-lite/list": "2",
     "@lerna-lite/version": "2",
     "@napi-rs/cli": "^2.11.0",
-    "@swc-node/register": "^1.9.0",
-    "@swc/core": "^1.4.12",
-    "@swc/jest": "^0.2.36",
+    "@swc-node/register": "^1.5.4",
+    "@swc/core": "^1.3.56",
+    "@swc/jest": "^0.2.23",
     "@taplo/cli": "^0.5.0",
     "@testing-library/jest-dom": "^6.0.0",
     "@testing-library/react": "^14.0.0",
@@ -101,7 +101,7 @@
     "http-server": "^14.0.0",
     "husky": "^8.0.0",
     "jasmine-expect": "^5.0.0",
-    "jest": "^29.7.0",
+    "jest": "29",
     "jest-environment-jsdom": "29",
     "json-stable-stringify": "^1.0.1",
     "karma": "^6.0.4",
@@ -123,8 +123,8 @@
     "rimraf": "^3.0.2",
     "serialize-javascript": "^6.0.0",
     "test262-harness": "10",
-    "ts-jest": "^29.1.2",
-    "ts-loader": "^9.5.1",
+    "ts-jest": "29",
+    "ts-loader": "^9.1.2",
     "tsd": "^0.30.0",
     "tslib": "^2.4.0",
     "typescript": "5.2.2",
@@ -145,12 +145,6 @@
     "source-map-js": "1.0.2"
   },
   "pnpm": {
-    "overrides": {
-      "ember-template-recast": "^6.1.4",
-      "@glimmer/reference": "^0.91.1",
-      "@glimmer/syntax": "^0.91.1",
-      "@glimmer/validator": "^0.91.1"
-    },
     "packageExtensions": {
       "ts-jest": {
         "dependencies": {
