+++ packages/cli-lib/tests/unit/gjs_extractor.test.ts
@@ -1,11 +1,11 @@
 import {MessageDescriptor, interpolateName} from '@formatjs/ts-transformer'
-import {parseFile} from '../../src/hbs_extractor'
+import {parseFile} from '../../src/gts_extractor'
 import {readFile} from 'fs-extra'
 import {join} from 'path'
 
-test('hbs_extractor', async function () {
+test('gts_extractor', async function () {
   let messages: MessageDescriptor[] = []
-  const fixturePath = join(__dirname, './fixtures/comp.hbs')
+  const fixturePath = join(__dirname, './fixtures/comp.gjs')
   parseFile(await readFile(fixturePath, 'utf8'), fixturePath, {
     onMsgExtracted(_: any, msgs: any) {
       messages = messages.concat(msgs)
