+++ packages/swc-plugin/package.json
@@ -19,7 +19,7 @@
   "dependencies": {
     "@formatjs/icu-messageformat-parser": "workspace:*",
     "@formatjs/ts-transformer": "workspace:*",
-    "@swc/core": "1.4.12",
+    "@swc/core": "1.2.241",
     "@types/json-stable-stringify": "^1.0.32",
     "@types/node": "14 || 16 || 17",
     "chalk": "^4.0.0",
